Fixed a problem that can lead to assertion error if any paired-end read's insert size > 32767 (by changing the type of insertL in PairedEndHit.h from short to int)

+++ PairedEndHit.h
@@ -15,13 +15,13 @@ public:
 		this->insertL = insertL;
 	}
 
-	short getInsertL() const { return insertL; }
+	int getInsertL() const { return insertL; }
 
 	bool read(std::istream&);
 	void write(std::ostream&);
 
 private:
-	short insertL; // insert length
+	int insertL; // insert length
 };
 
 bool PairedEndHit::read(std::istream& in) {
